test: split claude-base reaper test to satisfy file line limit (#509)
The agent's new `test_build_worker_runtime_wires_claude_base_reaper` pushed
`tests/unit/service/test_worker.py` to 1504 lines, tripping the
`test_first_party_code_files_stay_under_line_limit` guardrail (max 1500).

Move that single test into a sibling module
`tests/unit/service/test_worker_claude_base_reaper.py`, importing the shared
`_settings` / `_in_process_merge_coordinator` builders from the parent module
(matching the existing cross-module test-helper import convention). The test
body is unchanged, so coverage is identical; test_worker.py drops to 1410 lines.
